Prevent browser caching off all files
Unfortunately S3 only offers "project wide" response headers config. We
need to make sure that *.html files are not cached on S3, as these
settings are passed along to cloudfront.

Since we have no other mechanism of controlling cache we do it here. In
the future I'd like to split the deploy into two parts: assets with a
far future expires header, and documents with no caching.